Adding numbet of users hook
